requirements.txt
appdirs==1.4.0
decorator==4.0.11
ipython==5.1.0
ipython-genutils==0.1.0
Keras==2.0.8
numpy==1.12.0
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.9
protobuf==3.1.0.post1
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
PyYAML==3.12
scipy==0.18.1
simplegeneric==0.8.1
six==1.10.0
tensorflow==1.13.0
traitlets==4.3.1
wcwidth==0.1.7
